safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/style-blocked.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="style-src 'none'; script-src 'unsafe-inline'">
<link rel="stylesheet" href="resources/blue.css">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<script>
document.write(document.styleSheets.length > 0 ? 'FAIL' : 'PASS');
</script>
</body>
</html>
