safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/style-blocked-expected.txt

CONSOLE MESSAGE: Refused to load the stylesheet 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/blue.css' because it violates the following Content Security Policy directive: "style-src 'none'".

PASS
